Move service to camera
